Renaming PhpSpec Adapter
